Move upstream integration service into its own folder
