+++ apps/desktop/src/components/IntegrateUpstreamModal.svelte
@@ -4,9 +4,6 @@
 	import SelectItem from '$components/SelectItem.svelte';
 	import { BaseBranchService } from '$lib/baseBranch/baseBranchService';
 	import { getForge } from '$lib/forge/interface/forge';
-	import { copyToClipboard } from '$lib/utils/clipboard';
-	import { openExternalUrl } from '$lib/utils/url';
-	import { BranchStack } from '$lib/vbranches/types';
 	import {
 		getBaseBrancheResolution,
 		getResolutionApproach,
@@ -19,7 +16,10 @@
 		type StackStatus,
 		stackFullyIntegrated,
 		type BranchStatus
-	} from '$lib/vbranches/upstreamIntegrationService';
+	} from '$lib/upstream/upstreamIntegrationService';
+	import { copyToClipboard } from '$lib/utils/clipboard';
+	import { openExternalUrl } from '$lib/utils/url';
+	import { BranchStack } from '$lib/vbranches/types';
 	import { getContext } from '@gitbutler/shared/context';
 	import Badge from '@gitbutler/ui/Badge.svelte';
 	import Button from '@gitbutler/ui/Button.svelte';
